script: Move script_thread and related structs to `script/event_loop`
Testing: it compiles
