+++ components/script/dom/bindings/trace.rs
@@ -48,8 +48,8 @@ use crate::dom::bindings::refcounted::{Trusted, TrustedPromise};
 use crate::dom::html::htmlmediaelement::HTMLMediaElementFetchContext;
 use crate::dom::srcset::SourceSet;
 use crate::dom::windowproxy::WindowProxyHandler;
+use crate::event_loop::script_thread::IncompleteParserContexts;
 use crate::script_runtime::StreamConsumer;
-use crate::script_thread::IncompleteParserContexts;
 use crate::tasks::task::TaskBox;
 
 /// Wrapper type for nop traceble
