+++ components/script/dom/document/document.rs
@@ -80,7 +80,6 @@ use time::Duration as TimeDuration;
 use url::{Host, Position};
 
 use crate::animations::Animations;
-use crate::document_loader::{DocumentLoader, LoadType};
 use crate::dom::FlatTreeParent;
 use crate::dom::animationtimeline::AnimationTimeline;
 use crate::dom::attr::Attr;
@@ -205,13 +204,14 @@ use crate::dom::window::scrolling_box::{ScrollAxisState, ScrollingBox};
 use crate::dom::windowproxy::WindowProxy;
 use crate::dom::xpathevaluator::XPathEvaluator;
 use crate::dom::xpathexpression::XPathExpression;
+use crate::event_loop::document_loader::{DocumentLoader, LoadType};
+use crate::event_loop::script_thread::{ScriptThread, SharedRwLocks};
 use crate::fetch::{DeferredFetchRecordInvokeState, FetchCanceller};
 use crate::iframe_collection::IFrameCollection;
 use crate::image_animation::ImageAnimationManager;
 use crate::mime::{APPLICATION, CHARSET};
 use crate::navigation::navigate;
 use crate::network_listener::{FetchResponseListener, NetworkListener};
-use crate::script_thread::{ScriptThread, SharedRwLocks};
 use crate::stylesheet_loader::StylesheetContextId;
 use crate::stylesheet_set::StylesheetSetRef;
 use crate::tasks::task::NonSendTaskBox;
