+++ components/script/dom/domparser.rs
@@ -7,7 +7,6 @@ use js::rust::HandleObject;
 use script_bindings::reflector::{Reflector, reflect_dom_object_with_proto};
 use script_traits::DocumentActivity;
 
-use crate::document_loader::DocumentLoader;
 use crate::dom::bindings::codegen::Bindings::DOMParserBinding;
 use crate::dom::bindings::codegen::Bindings::DOMParserBinding::DOMParserMethods;
 use crate::dom::bindings::codegen::Bindings::DOMParserBinding::SupportedType::{
@@ -22,6 +21,7 @@ use crate::dom::document::{Document, DocumentSource, HasBrowsingContext, IsHTMLD
 use crate::dom::servoparser::ServoParser;
 use crate::dom::trustedtypes::trustedhtml::TrustedHTML;
 use crate::dom::window::Window;
+use crate::event_loop::document_loader::DocumentLoader;
 
 #[dom_struct]
 pub(crate) struct DOMParser {
