+++ components/script/dom/html/embedded_content/htmltrackelement.rs
@@ -42,11 +42,11 @@ use crate::dom::texttrack::TextTrack;
 use crate::dom::virtualmethods::VirtualMethods;
 use crate::dom::webvtt::vttcue::VTTCue;
 use crate::dom::{AttributeMutation, cors_setting_for_element};
+use crate::event_loop::script_thread::ScriptThread;
 use crate::fetch::{RequestWithGlobalScope, create_a_potential_cors_request};
 use crate::microtask::MicrotaskRunnable;
-use crate::network_listener::{FetchResponseListener, ResourceTimingListener};
+use crate::network_listener::{self, FetchResponseListener, ResourceTimingListener};
 use crate::realms::enter_auto_realm;
-use crate::{ScriptThread, network_listener};
 
 #[derive(Clone, Copy, Default, JSTraceable, MallocSizeOf, PartialEq)]
 #[repr(u16)]
