+++ components/script/dom/promise/promise.rs
@@ -43,9 +43,9 @@ use crate::dom::bindings::reflector::DomGlobal;
 use crate::dom::bindings::root::{AsHandleValue, Dom};
 use crate::dom::globalscope::GlobalScope;
 use crate::dom::promisenativehandler::{Callback, PromiseNativeHandler};
+use crate::event_loop::script_thread::ScriptThread;
 use crate::microtask::MicrotaskRunnable;
 use crate::realms::enter_auto_realm;
-use crate::script_thread::ScriptThread;
 
 #[dom_struct]
 #[cfg_attr(crown, crown::unrooted_must_root_lint::allow_unrooted_in_rc)]
