+++ components/script/event_loop/script_thread.rs
@@ -112,8 +112,6 @@ use url::Position;
 use webgpu_traits::{WebGPUDevice, WebGPUMsg};
 
 use crate::devtools::DevtoolsState;
-use crate::document_collection::DocumentCollection;
-use crate::document_loader::DocumentLoader;
 use crate::dom::bindings::codegen::Bindings::DocumentBinding::{
     DocumentMethods, DocumentReadyState,
 };
@@ -144,6 +142,10 @@ use crate::dom::types::DebuggerGlobalScope;
 use crate::dom::webgpu::identityhub::IdentityHub;
 use crate::dom::window::Window;
 use crate::dom::windowproxy::{CreatorBrowsingContextInfo, WindowProxy};
+use crate::event_loop::document_collection::DocumentCollection;
+use crate::event_loop::document_loader::DocumentLoader;
+use crate::event_loop::script_mutation_observers::ScriptMutationObservers;
+use crate::event_loop::script_window_proxies::ScriptWindowProxies;
 use crate::fetch::FetchCanceller;
 use crate::messaging::{
     CommonScriptMsg, MainThreadScriptMsg, MixedMessage, ScriptEventLoopSender,
@@ -154,11 +156,9 @@ use crate::mime::{APPLICATION, CHARSET, MimeExt, TEXT, XML};
 use crate::navigation::{InProgressLoad, NavigationListener};
 use crate::network_listener::{FetchResponseListener, submit_timing};
 use crate::realms::enter_auto_realm;
-use crate::script_mutation_observers::ScriptMutationObservers;
 use crate::script_runtime::{
     IntroductionType, Runtime, ScriptThreadEventCategory, ThreadSafeJSContext, get_reports,
 };
-use crate::script_window_proxies::ScriptWindowProxies;
 use crate::svg_font::SvgFontResolver;
 use crate::tasks::task_queue::TaskQueue;
 use crate::webdriver_handlers::jsval_to_webdriver;
