+++ components/script/lib.rs
@@ -22,23 +22,21 @@ extern crate stylo_atoms;
 
 mod animations;
 mod css;
-mod script_window_proxies;
 #[macro_use]
 mod tasks;
 mod body;
 pub(crate) mod conversions;
 mod devtools;
-pub(crate) mod document_loader;
 #[macro_use]
 mod dom;
 pub(crate) use dom::canvas_context;
+mod event_loop;
 pub(crate) mod fetch;
 pub(crate) mod indexeddb;
 mod init;
 mod layout_image;
 mod url;
 
-pub(crate) mod document_collection;
 pub(crate) mod iframe_collection;
 pub(crate) mod image_animation;
 pub mod layout_dom;
@@ -51,10 +49,7 @@ mod navigation;
 mod network_listener;
 mod realms;
 mod routed_promise;
-mod script_mutation_observers;
 pub(crate) mod script_runtime;
-#[expect(unsafe_code)]
-pub(crate) mod script_thread;
 pub(crate) mod serviceworker_manager;
 mod stylesheet_loader;
 mod stylesheet_set;
@@ -70,11 +65,11 @@ mod unminify;
 mod drag_data_store;
 mod links;
 
+pub use event_loop::script_thread::ScriptThread;
 pub use init::init;
 pub(crate) use script_bindings::DomTypes;
 pub(crate) use script_bindings::reflector::{AssociatedMemory, DomObject, MutDomObject, Reflector};
 pub use script_runtime::JSEngineSetup;
-pub use script_thread::ScriptThread;
 pub use serviceworker_manager::ServiceWorkerManager;
 
 pub(crate) use crate::dom::bindings::codegen::DomTypeHolder::DomTypeHolder;
